% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createCrisisVariables}
\alias{createCrisisVariables}
\title{Create Crisis Variables}
\usage{
createCrisisVariables(crisisDT, crisisCol = "Banking Crisis",
  idCol = "iso3", timeCol = "date")
}
\arguments{
\item{crisisDT}{}

\item{crisisCol}{}

\item{idCol}{}

\item{timeCol}{}
}
\value{
Original data.table with included countdown variables.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

