% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eventCounter}
\alias{eventCounter}
\title{Crisis Event Counter}
\usage{
eventCounter(x, success = 0, failure = 1)
}
\arguments{
\item{x}{}

\item{success}{}

\item{failure}{}
}
\value{
A vector with event enumeration.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

