% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBloombergSovCDS}
\alias{getBloombergSovCDS}
\title{Create a dataset of Sovereign CDS Spreads (source: Bloomberg)}
\usage{
getBloombergSovCDS()
}
\value{
data.table with cds
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

