% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIMFIFS}
\alias{getIMFIFS}
\title{Create a IMF-IFS dataset}
\usage{
getIMFIFS()
}
\value{
data.table with resulting dataset (attribute 'colnames') contains
labeling information of variables.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

