% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSovBenchmarks}
\alias{getSovBenchmarks}
\title{Prepare a dataset of sovereign benchmark credit risk measures}
\usage{
getSovBenchmarks()
}
\value{
data.table with the resulting data set
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

