% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSovBondSpreads}
\alias{getSovBondSpreads}
\title{Get a dataset of sovereign bond spreads (source: World Bank and IMF)}
\usage{
getSovBondSpreads()
}
\value{
data.table of results
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

