% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSPRatings}
\alias{getSPRatings}
\title{Create a dataset of S&P sovereign credit ratings}
\usage{
getSPRatings()
}
\value{
data.table with ratings
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

