% Generated by roxygen2 (4.0.2): do not edit by hand
\name{loadCrisisDB}
\alias{loadCrisisDB}
\title{Load Reinhardt and Rogoff Sovereign Crisis Database}
\usage{
loadCrisisDB()
}
\value{
data.table with crisis dataset.
}
\description{
This function returns a dataset of Reinhardt and Rogoff, which dates 6
different types of crises for a set of 70 countries.
}
\details{
TODO: RR's database ends in 2010. Update the DB by including the sovereign
debt crisis episodes during 2010-2014.
}
\author{
Janko Cizel
}

