% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotSovBenchmarks}
\alias{plotSovBenchmarks}
\title{Line plots of sovereign credit ratings, CDS spreads, and government
bond spreads}
\usage{
plotSovBenchmarks(isoSel = "ARG", limits = as.Date(c("1990-01-01",
  "2013-01-01")), filename = "~/Downloads/test.pdf")
}
\arguments{
\item{isoSel}{ISO3 country code of a country for which the plot is to be created.}

\item{limits}{a Date vector indicating the time range of the plot}

\item{filename}{path of the resulting plot (in pdf format)}
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

