% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summarizeDataAvailability}
\alias{summarizeDataAvailability}
\title{Summarize data availability for each cross-sectional unit over time}
\usage{
summarizeDataAvailability(dt, idCol = "iso3", timeCol = "date")
}
\arguments{
\item{dt}{data.table containing the data}

\item{idCol}{column with cross-sectional unit identifiers}

\item{timeCol}{column with time-unit identifiers}
}
\value{
data.table with a summary of data availability
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

