% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tabulateCrises}
\alias{tabulateCrises}
\title{Produce a Summary Table of Sovereign Crises}
\usage{
tabulateCrises(crisisDT, crisisTypes = "Foreign Sov Debt", timeCol = "YEAR",
  min.time = 1995, idCol = "iso3", outfile)
}
\arguments{
\item{crisisTypes}{}

\item{timeCol}{}

\item{min.time}{}

\item{idCol}{}

\item{crisisDB}{}
}
\value{
Latex table (without table environment)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Janko Cizel
}

